Move chai matchers to their own package
