+++ packages/hardhat-chai-matchers/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  extends: [`${__dirname}/../../config/eslint/eslintrc.js`],
+  parserOptions: {
+    project: `${__dirname}/tsconfig.json`,
+    sourceType: "module",
+  },
+};
