+++ packages/hardhat-chai-matchers/.mocharc.json
@@ -1,0 +1,5 @@
+{
+  "require": "ts-node/register/files",
+  "ignore": ["test/fixture-projects/**/*"],
+  "timeout": 10000
+}
