+++ packages/hardhat-chai-matchers/.prettierignore
@@ -1,0 +1,12 @@
+/node_modules
+/dist
+/internal
+/types
+/*.d.ts
+/*.d.ts.map
+/*.js
+/*.js.map
+/build-test
+/test/fixture-projects/**/artifacts
+/test/fixture-projects/**/cache
+CHANGELOG.md
