+++ packages/hardhat-chai-matchers/README.md
@@ -1,0 +1,3 @@
+[![npm](https://img.shields.io/npm/v/@nomiclabs/hardhat-chai-matchers.svg)](https://www.npmjs.com/package/@nomiclabs/hardhat-chai-matchers)
+
+# hardhat-chai-matchers
