+++ packages/hardhat-chai-matchers/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@nomiclabs/hardhat-test-helpers",
+  "name": "@nomiclabs/hardhat-chai-matchers",
   "version": "0.0.1",
   "description": "Hardhat utils for testing",
-  "homepage": "https://github.com/nomiclabs/hardhat/tree/master/packages/hardhat-test-helpers",
+  "homepage": "https://github.com/nomiclabs/hardhat/tree/master/packages/hardhat-chai-matchers",
   "repository": "github:nomiclabs/hardhat",
   "author": "Nomic Labs LLC",
   "license": "MIT",
@@ -19,7 +19,7 @@
     "lint:fix": "yarn prettier --write && yarn eslint --fix",
     "eslint": "eslint 'src/**/*.ts' 'test/**/*.ts'",
     "prettier": "prettier \"**/*.{js,md,json}\"",
-    "test": "mocha --recursive \"test/**/*.ts\" --exit",
+    "test": "mocha --recursive \"test/**/*.ts\" --exit --reporter dot",
     "build": "tsc --build .",
     "clean": "rimraf dist internal types *.{d.ts,js}{,.map} build-test tsconfig.tsbuildinfo"
   },
@@ -36,16 +36,19 @@
     "README.md"
   ],
   "devDependencies": {
+    "@types/bn.js": "^5.1.0",
     "@types/chai": "^4.2.0",
     "@types/mocha": "^9.0.0",
     "@types/node": "^12.0.0",
     "@typescript-eslint/eslint-plugin": "4.29.2",
     "@typescript-eslint/parser": "4.29.2",
-    "chai": "^4.2.0",
+    "bignumber.js": "^9.0.2",
+    "bn.js": "^5.1.0",
     "eslint": "^7.29.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-import": "2.24.1",
     "eslint-plugin-prettier": "3.4.0",
+    "ethers": "^5.0.0",
     "hardhat": "^2.0.0",
     "mocha": "^7.2.0",
     "prettier": "2.4.1",
@@ -54,6 +57,10 @@
     "typescript": "~4.5.2"
   },
   "peerDependencies": {
+    "ethers": "^5.0.0",
     "hardhat": "^2.0.0"
+  },
+  "dependencies": {
+    "chai": "^4.2.0"
   }
 }
