+++ packages/hardhat-chai-matchers/src/bnChai.ts
@@ -1,4 +1,4 @@
-import "../types";
+import "./types";
 import { supportBigNumber } from "./bigNumber";
 
 export function bnChai(chai: Chai.ChaiStatic, utils: Chai.ChaiUtils) {
