+++ packages/hardhat-chai-matchers/src/index.ts
@@ -1,0 +1,1 @@
+export { bnChai } from "./bnChai";
