+++ packages/hardhat-chai-matchers/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "../../config/typescript/tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./dist"
+  },
+  "exclude": ["./dist", "./node_modules", "./test/**/hardhat.config.ts"],
+  "references": [
+    {
+      "path": "../hardhat-core/src"
+    }
+  ]
+}
