+++ packages/hardhat-test-helpers/src/index.ts
@@ -1,2 +1,1 @@
 export * from "./setupContracts";
-export { bnChai } from "./chai/matchers/bnChai";
